all/mysite/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
channels==4.0.0
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==4.1.5
django-channels==0.7.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
fonttools==4.39.0
fpdf2==2.6.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
oauthlib==3.2.2
Pillow==9.4.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
